backend/requirements.txt
Flask
Flask-Cors
requests
gunicorn
python-dotenv
PyPDF2
python-docx
firebase-admin
